Fix imports of refactored classes
