+++ backend/src/main/java/de/neuefische/greensitter/controller/ApiSearchController.java
@@ -1,7 +1,8 @@
 package de.neuefische.greensitter.controller;
 
 import de.neuefische.greensitter.api.ApiSearchService;
-import de.neuefische.greensitter.api.PlantData;
+import de.neuefische.greensitter.api.dtos.SearchResultData;
+import de.neuefische.greensitter.service.PlantService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -14,12 +15,13 @@ public class ApiSearchController {
 
     private final ApiSearchService searchService;
 
-    public ApiSearchController(ApiSearchService searchService) {
+
+    public ApiSearchController(ApiSearchService searchService, PlantService plantService) {
         this.searchService = searchService;
     }
 
     @GetMapping
-    public PlantData[] getQueryResults(@RequestParam String query){
+    public SearchResultData[] getQueryResults(@RequestParam String query){
         return searchService.getSearchResults(query);
     }
 }
